Moved lcos calcs to common_financial from common
