+++ ssc/common.cpp
@@ -130,372 +130,349 @@ var_info vtab_oandm[] = {
 { SSC_INPUT,SSC_ARRAY   , "om_fuelcell_capacity_cost"                         , "Fuel cell capacity-based System Costs amount"                   , "$/kWcap"                                , ""                                      , "System Costs"         , "?=0.0"          , ""                      , ""},
 var_info_invalid };
 
-var_info vtab_lcos_inputs[] = {
-    /*   VARTYPE           DATATYPE         NAME                             LABEL                                UNITS      META                 GROUP          REQUIRED_IF                 CONSTRAINTS                      UI_HINTS*/
+//var_info vtab_lcos_inputs[] = {
+//    /*   VARTYPE           DATATYPE         NAME                             LABEL                                UNITS      META                 GROUP          REQUIRED_IF                 CONSTRAINTS                      UI_HINTS*/
+//
+//    { SSC_INPUT,        SSC_ARRAY,      "batt_annual_charge_from_system",                 "Battery annual energy charged from system",                 "kWh",      "",                      "Battery",       "",                           "",                               "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "batt_annual_discharge_energy",               "Battery annual energy discharged",                      "kWh",      "",                      "Battery",       "",                           "",                               "" },
+//    { SSC_INPUT,        SSC_NUMBER,     "batt_salvage_percentage",                     "Net pre-tax cash battery salvage value",	                               "%",	 "",					  "Financial Parameters",             "?=0",                     "MIN=0,MAX=100",      			"" },
+//
+//    { SSC_INPUT,        SSC_NUMBER,      "battery_total_cost_lcos",               "Battery total investment cost",                      "$",      "",                      "Battery",       "",                           "",                               "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "grid_to_batt",                               "Electricity to grid from battery",                      "kW",      "",                       "Battery",       "",                           "",                              "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "buy_rate_ts", "TOU buy rate for energy charges (year 1 hourly)", "", "", "Time Series", "", "", "" },
+//    { SSC_INPUT, SSC_ARRAY, "year1_monthly_ec_charge_with_system", "Energy charge with system", "$", "", "Charges by Month", "", "", "" },
+//    { SSC_INPUT, SSC_ARRAY, "year1_monthly_ec_charge_gross_with_system", "Energy charge with system before credits", "$/mo", "", "Monthly", "", "LENGTH=12", "" },
+//    { SSC_INPUT,       SSC_ARRAY,      "year1_monthly_electricity_to_grid",    "Electricity to/from grid",           "kWh/mo", "", "Monthly",          "",                         "LENGTH=12",                     "" },
+//    { SSC_INPUT, SSC_MATRIX, "charge_w_sys_ec_ym", "Energy charge with system", "$", "", "Charges by Month", "", "", "COL_LABEL=MONTHS,FORMAT_SPEC=CURRENCY,GROUP=UR_AM" },
+//    { SSC_INPUT, SSC_MATRIX, "true_up_credits_ym",     "Net annual true-up payments", "$", "", "Charges by Month", "", "", "COL_LABEL=MONTHS,FORMAT_SPEC=CURRENCY,GROUP=UR_AM" },
+//    { SSC_INPUT,        SSC_ARRAY,      "batt_capacity_percent",                      "Battery relative capacity to nameplate",                 "%",        "",                     "Battery",       "",                           "",                              "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "monthly_grid_to_batt",                       "Energy to battery from grid",                           "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "monthly_batt_to_grid",                       "Energy to grid from battery",                           "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "monthly_grid_to_load",                       "Energy to load from grid",                              "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
+//    { SSC_INPUT,        SSC_ARRAY,      "monthly_system_to_grid",                     "Energy to grid from system",                            "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_cost_lcos", "Annual storage costs", "$", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_discharge_lcos", "Annual storage discharge", "kWh", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid", "Annual cost to charge from grid", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid_month", "Annual cost to charge from grid (monthly)", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_pv", "Annual cost to charge from system", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_production_expense", "Annual cost to for battery production based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_capacity_expense", "Annual cost for battery capacity based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_fixed_expense", "Annual fixed cost for battery maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_batt_replacement_cost", "Annual cost of battery replacements", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//    { SSC_OUTPUT, SSC_ARRAY, "cf_salvage_cost_lcos", "Annual battery salvage value costs", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
+//
+//    { SSC_OUTPUT, SSC_NUMBER, "npv_annual_costs_lcos", "Present value of annual storage costs", "$", "", "LCOE calculations", "", "", "" },
+//    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_real", "Present value of annual stored energy (real)", "kWh", "", "LCOE calculations", "", "", "" },
+//    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_nom",                        "Nominal levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
+//    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_real",                        "Real levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
+//    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_nom", "Present value of annual stored energy (nominal)", "kWh", "", "LCOE calculations", "", "", "" },
+//    var_info_invalid };
 
-    { SSC_INPUT,        SSC_ARRAY,      "batt_annual_charge_from_system",                 "Battery annual energy charged from system",                 "kWh",      "",                      "Battery",       "",                           "",                               "" },
-    { SSC_INPUT,        SSC_ARRAY,      "batt_annual_discharge_energy",               "Battery annual energy discharged",                      "kWh",      "",                      "Battery",       "",                           "",                               "" },
-    { SSC_INPUT,        SSC_NUMBER,     "batt_salvage_percentage",                     "Net pre-tax cash battery salvage value",	                               "%",	 "",					  "Financial Parameters",             "?=0",                     "MIN=0,MAX=100",      			"" },
-
-    { SSC_INPUT,        SSC_NUMBER,      "battery_total_cost_lcos",               "Battery total investment cost",                      "$",      "",                      "Battery",       "",                           "",                               "" },
-    { SSC_INPUT,        SSC_ARRAY,      "grid_to_batt",                               "Electricity to grid from battery",                      "kW",      "",                       "Battery",       "",                           "",                              "" },
-    { SSC_INPUT,        SSC_ARRAY,      "buy_rate_ts", "TOU buy rate for energy charges (year 1 hourly)", "", "", "Time Series", "", "", "" },
-    { SSC_INPUT, SSC_ARRAY, "year1_monthly_ec_charge_with_system", "Energy charge with system", "$", "", "Charges by Month", "", "", "" },
-    { SSC_INPUT, SSC_ARRAY, "year1_monthly_ec_charge_gross_with_system", "Energy charge with system before credits", "$/mo", "", "Monthly", "", "LENGTH=12", "" },
-    { SSC_INPUT,       SSC_ARRAY,      "year1_monthly_electricity_to_grid",    "Electricity to/from grid",           "kWh/mo", "", "Monthly",          "",                         "LENGTH=12",                     "" },
-    { SSC_INPUT, SSC_MATRIX, "charge_w_sys_ec_ym", "Energy charge with system", "$", "", "Charges by Month", "", "", "COL_LABEL=MONTHS,FORMAT_SPEC=CURRENCY,GROUP=UR_AM" },
-    { SSC_INPUT, SSC_MATRIX, "true_up_credits_ym",     "Net annual true-up payments", "$", "", "Charges by Month", "", "", "COL_LABEL=MONTHS,FORMAT_SPEC=CURRENCY,GROUP=UR_AM" },
-    { SSC_INPUT,        SSC_ARRAY,      "batt_capacity_percent",                      "Battery relative capacity to nameplate",                 "%",        "",                     "Battery",       "",                           "",                              "" },
-    { SSC_INPUT,        SSC_ARRAY,      "monthly_grid_to_batt",                       "Energy to battery from grid",                           "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
-    { SSC_INPUT,        SSC_ARRAY,      "monthly_batt_to_grid",                       "Energy to grid from battery",                           "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
-    { SSC_INPUT,        SSC_ARRAY,      "monthly_grid_to_load",                       "Energy to load from grid",                              "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
-    { SSC_INPUT,        SSC_ARRAY,      "monthly_system_to_grid",                     "Energy to grid from system",                            "kWh",      "",                      "Battery",       "",                          "LENGTH=12",                     "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_cost_lcos", "Annual storage costs", "$", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_discharge_lcos", "Annual storage discharge", "kWh", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid", "Annual cost to charge from grid", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid_month", "Annual cost to charge from grid (monthly)", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_pv", "Annual cost to charge from system", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_production_expense", "Annual cost to for battery production based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_capacity_expense", "Annual cost for battery capacity based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_fixed_expense", "Annual fixed cost for battery maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_batt_replacement_cost", "Annual cost of battery replacements", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_salvage_cost_lcos", "Annual battery salvage value costs", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-
-    { SSC_OUTPUT, SSC_NUMBER, "npv_annual_costs_lcos", "Present value of annual storage costs", "$", "", "LCOE calculations", "", "", "" },
-    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_real", "Present value of annual stored energy (real)", "kWh", "", "LCOE calculations", "", "", "" },
-    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_nom",                        "Nominal levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
-    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_real",                        "Real levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
-    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_nom", "Present value of annual stored energy (nominal)", "kWh", "", "LCOE calculations", "", "", "" },
-    var_info_invalid };
-
-var_info vtab_lcos_outputs[] = {
-    /*   VARTYPE           DATATYPE         NAME                             LABEL                                UNITS      META                 GROUP          REQUIRED_IF                 CONSTRAINTS                      UI_HINTS*/
-
-    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_cost_lcos", "Annual storage costs", "$", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_annual_discharge_lcos", "Annual storage discharge", "kWh", "", "LCOS calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid", "Annual cost to charge from grid", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_grid_month", "Annual cost to charge from grid (monthly)", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-
-    { SSC_OUTPUT, SSC_ARRAY, "cf_charging_cost_pv", "Annual cost to charge from system", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_production_expense", "Annual cost to for battery production based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_capacity_expense", "Annual cost for battery capacity based maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_om_batt_fixed_expense", "Annual fixed cost for battery maintenance", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_batt_replacement_cost", "Annual cost of battery replacements", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-    { SSC_OUTPUT, SSC_ARRAY, "cf_salvage_cost_lcos", "Annual battery salvage value costs", "$", "", "LCOE calculations", "", "LENGTH_EQUAL=cf_length", "" },
-
-    { SSC_OUTPUT, SSC_NUMBER, "npv_annual_costs_lcos", "Present value of annual storage costs", "$", "", "LCOE calculations", "", "", "" },
-    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_real", "Present value of annual stored energy (real)", "kWh", "", "LCOE calculations", "", "", "" },
-    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_nom",                        "Nominal levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
-    { SSC_OUTPUT,       SSC_NUMBER,     "lcos_real",                        "Real levelized cost of storage",              "cents/kWh",                   "", "Metrics", "", "", "" },
-
-    { SSC_OUTPUT, SSC_NUMBER, "npv_energy_lcos_nom", "Present value of annual stored energy (nominal)", "kWh", "", "LCOE calculations", "", "", "" },
-    var_info_invalid };
-
-double npv(int cf_line, int nyears, double rate, util::matrix_t<double> cf)
-{
-    //if (rate == -1.0) throw general_error("cannot calculate NPV with discount rate equal to -1.0");
-    double rr = 1.0;
-    if (rate != -1.0) rr = 1.0 / (1.0 + rate);
-    double result = 0;
-    for (int i = nyears; i > 0; i--)
-        result = rr * result + cf.at(cf_line, i);
-
-    return result * rr;
-}
-
-void escal_or_annual(int cf_line, int nyears, const std::string& variable,
-    double inflation_rate, double scale,  util::matrix_t<double> cf, compute_module *cm, bool as_rate = true,  double escal = 0.0 )
-{
-    size_t count;
-    ssc_number_t* arrp = cm->as_array(variable, &count);
-
-    if (as_rate)
-    {
-        if (count == 1)
-        {
-            escal = inflation_rate + scale * arrp[0];
-            for (int i = 0; i < nyears; i++)
-                cf.at(cf_line, i + 1) = pow(1 + escal, i);
-        }
-        else
-        {
-            for (int i = 0; i < nyears && i < (int)count; i++)
-                cf.at(cf_line, i + 1) = 1 + arrp[i] * scale;
-        }
-    }
-    else
-    {
-        if (count == 1)
-        {
-            for (int i = 0; i < nyears; i++)
-                cf.at(cf_line, i + 1) = arrp[0] * scale * pow(1 + escal + inflation_rate, i);
-        }
-        else
-        {
-            for (int i = 0; i < nyears && i < (int)count; i++)
-                cf.at(cf_line, i + 1) = arrp[i] * scale;
-        }
-    }
-}
-
-void save_cf(int cf_line, int nyears, const std::string& name, util::matrix_t<double> cf, compute_module *cm)
-{
-    ssc_number_t* arrp = cm->allocate(name, nyears + 1);
-    for (int i = 0; i <= nyears; i++)
-        arrp[i] = (ssc_number_t)cf.at(cf_line, i);
-}
-
-void lcos_calc(compute_module* cm, util::matrix_t<double> cf, int nyears, double nom_discount_rate, double inflation_rate, double lcoe_real, double total_cost, double real_discount_rate, int grid_charging_cost_version, ssc_number_t* ppa_multipliers = { 0 }) {
-    enum {
-        CF_battery_replacement_cost_lcos,
-        CF_ppa_price_lcos,
-        CF_om_fixed_expense_lcos,
-        CF_om_production_expense_lcos,
-        CF_om_capacity_expense_lcos,
-        CF_om_fixed1_expense_lcos,
-        CF_om_production1_expense_lcos,
-        CF_om_capacity1_expense_lcos,
-        CF_om_fixed2_expense_lcos,
-        CF_om_production2_expense_lcos,
-        CF_om_capacity2_expense_lcos,
-        CF_om_fuel_expense_lcos,
-        CF_energy_charged_grid_lcos,
-        CF_energy_charged_pv_lcos,
-        CF_energy_discharged_lcos,
-        CF_charging_cost_pv_lcos,
-        CF_charging_cost_grid_lcos,
-        CF_om_cost_lcos_lcos,
-        CF_salvage_cost_lcos_lcos,
-        CF_investment_cost_lcos_lcos,
-        CF_annual_cost_lcos_lcos,
-        CF_util_escal_rate_lcos,
-        CF_degradation_lcos,
-
-    };
-
-    if (cm->is_assigned("battery_total_cost_lcos") && cm->as_double("battery_total_cost_lcos") != 0) { //Check that there are non-zero battery costs from the UI
-        double lcos_investment_cost = cm->as_double("battery_total_cost_lcos"); //Battery capital costs ($), does not include replacement costs
-        std::vector<double> charged_pv = cm->as_vector_double("batt_annual_charge_from_system"); //Annual kwh charged from PV or generic system (nyears array)
-        std::vector<double> charged_total = cm->as_vector_double("batt_annual_charge_energy"); //Total annual kwh charged by battery system (nyears array)
-        std::vector<double> lcos_energy_discharged = cm->as_vector_double("batt_annual_discharge_energy"); //Total annual kwh discharged by battery system (nyears array)
-        size_t n_grid_to_batt, n_monthly_grid_to_batt, n_monthly_grid_to_load, n_monthly_energy_charge, n_net_annual_true_up; //Preallocate size variables
-        ssc_number_t* grid_to_batt = cm->as_array("grid_to_batt", &n_grid_to_batt); //Power from grid to battery in kW hourly
-        ssc_number_t* monthly_grid_to_batt = cm->as_array("monthly_grid_to_batt", &n_monthly_grid_to_batt);
-        
-        ssc_number_t* monthly_batt_to_grid;
-        if (grid_charging_cost_version != 0) //For FOM systems the battery exports to grid rather than to load
-            monthly_batt_to_grid = cm->as_array("monthly_batt_to_grid", &n_monthly_grid_to_load); //monthly battery exports to grid for 1st year (kWh)
-        ssc_number_t* monthly_grid_to_load;
-        if(grid_charging_cost_version == 0) //For BTM systems the battery exports to load
-            monthly_grid_to_load = cm->as_array("monthly_grid_to_load", &n_monthly_grid_to_load);//monthly battery exports to load for first year (kWh)
-        
-
-        util::matrix_t<double> monthly_energy_charge; //monthly energy charges at 12 month x nyears matrix ($)
-        util::matrix_t<double> net_annual_true_up; //net annual true up payments as 12 month x nyears matrix ($)
-        size_t n_steps_per_year = 8760; //Initialize number of timesteps per year (calculated based on lifetime choice and nstep of weather file
-        //ssc_number_t* monthly_batt_to_grid = as_array("monthly_batt_to_grid", &n_monthly_grid_to_load);
-        ssc_number_t* monthly_system_to_grid = cm->as_array("monthly_system_to_grid", &n_monthly_grid_to_load); //monthly grid energy bought to satisfy load for first year (kWh)
-        ssc_number_t* monthly_electricity_tofrom_grid;
-        bool ppa_purchases = !(cm->is_assigned("en_electricity_rates") && cm->as_number("en_electricity_rates") == 1); //Does the system use ppa purchases in revenue calculations (Single Owner only)
-        if (!ppa_purchases || (grid_charging_cost_version == 0)) { //Are ppa purchases used in Single Owner or is the case using BTM financial model?
-            monthly_electricity_tofrom_grid = cm->as_array("year1_monthly_electricity_to_grid", &n_monthly_grid_to_load); //Monthly electricity exported to the grid in the first year (kWh)
-            monthly_energy_charge = cm->as_matrix("charge_w_sys_ec_ym"); //Use monthly energy charges from utility bill ($)
-            net_annual_true_up = cm->as_matrix("true_up_credits_ym"); //Use net annual true up payments regardless of billing mode ($)
-        }
-        
-        size_t n_mp_market_price;
-        ssc_number_t* mp_market_price;
-        if (grid_charging_cost_version == 3) //Only if financial model is Merchant Plant
-            mp_market_price = cm->as_array("mp_energy_market_price", &n_mp_market_price); //Energy market price (hourly) for lifetime array to calculate grid charging cost for Merchant Plant ($)
-
-        size_t n_degradation;
-        ssc_number_t* degradation = cm->as_array("degradation", &n_degradation);
-        if (n_degradation == 1)
-        {
-            for (size_t i = 1; i <= nyears; i++) cf.at(CF_degradation_lcos, i) = pow((1.0 - degradation[0] / 100.0), i - 1);
-        }
-        else if (n_degradation > 0)
-        {
-            for (size_t i = 0; i < nyears && i < (int)n_degradation; i++) cf.at(CF_degradation_lcos, i + 1) = (1.0 - degradation[i] / 100.0);
-        }
-
-        if (cm->as_integer("system_use_lifetime_output") == 1)
-            n_steps_per_year = n_grid_to_batt / nyears; //Steps per year = Lifetime array size / number years
-        else
-            n_steps_per_year = n_grid_to_batt; //If no lifetime, steps per year = first year array size
-        //std::vector<double> grid_to_batt = as_vector_double("grid_to_batt");
-        cf.at(CF_charging_cost_grid_lcos, 0) = 0; //Initialize year 0 charging cost to $0
-        
-
-        if (cm->is_assigned("rate_escalation")) //Create rate escalation nyears array with inflation and specified rate escalation %
-            escal_or_annual(CF_util_escal_rate_lcos, nyears, "rate_escalation", inflation_rate, 0.01, cf, cm, true, 0);
-        save_cf(CF_util_escal_rate_lcos, nyears, "cf_util_escal_rate", cf, cm);
-
-        double capex_lcoe_ratio = 1 / 0.8; //ratio of capex ratio between PV+batt / PV to LCOE ratio PV+batt/ PV (assumed based on table)
-        double lcoe_real_lcos = lcoe_real * capex_lcoe_ratio * (total_cost - lcos_investment_cost) / total_cost; //cents/kWh
-        //Using ration of investment cost to lcoe for system+storage and system only to approximate lcoe for system only; Used in PV charging cost calculations
-       
-
-        for (int a = 0; a <= nyears; a++) { //Iterate through nyears of the project
-
-
-            if (grid_charging_cost_version == 0) { //0 - Cashloan, BTM
-
-
-                for (size_t m = 0; m < 12; m++) { //monthly iteration for each year
-                    if (a != 0 && (monthly_grid_to_load[m] + monthly_grid_to_batt[m]) != 0) {
-                        //cf.at(CF_charging_cost_grid_month, a) += monthly_grid_to_batt[m] / (monthly_grid_to_batt[m] + monthly_grid_to_load[m]) * monthly_energy_charge[m] * charged_grid[a] / charged_grid[1] * cf.at(CF_util_escal_rate, a);
-                        cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / (monthly_grid_to_load[m] + monthly_grid_to_batt[m]) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //use the electricity rate data by year (also trueup) //* charged_grid[a] / charged_grid[1] * cf.at(CF_util_escal_rate, a);
-                    }
-                }
-            }
-            else if (grid_charging_cost_version == 1) { //Single Owner
-                ppa_purchases = !(cm->is_assigned("en_electricity_rates") && cm->as_number("en_electricity_rates") == 1); //Are ppa purchases used or not?
-                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
-                {
-                    //Lifetime calculations
-                    //Calculate cost to charge battery from grid (either using ppa price or electricity rates if enabled)
-                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a
-                    if (ppa_purchases && a != 0) {
-                        for (size_t h = 0; h < n_steps_per_year; h++) {
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(size_t(a) - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from PPA price ($)
-                        }
-                    }
-                    else if (!ppa_purchases && a != 0) {
-                        for (size_t m = 0; m < 12; m++) {
-                            if (((monthly_batt_to_grid[m] + monthly_system_to_grid[m]) + -monthly_electricity_tofrom_grid[m]) != 0)
-                                cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / ((monthly_batt_to_grid[m] + monthly_system_to_grid[m]) + -monthly_electricity_tofrom_grid[m]) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //Grid charging cost from monthly energy charges and net annual true up ($
-
-                        }
-                    }
-
-
-                }
-                else //Not lifetime
-                {
-                    //Single year calculations (might not be needed as their is no Single Owner for PVWatts Battery
-                    //Calculate cost to charge battery from grid (either using ppa price or electricity rates if enabled)
-                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a ($)
-
-                    if (ppa_purchases && a != 0) { //PPA purchases enabled and not in investment year
-                        for (size_t h = 0; h < 8760; h++) {
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost calculated from PPA price ($)
-                        }
-                    }
-                    else if (!ppa_purchases && a != 0) { //No PPA purchases and not in investment year
-                        for (size_t m = 0; m < 12; m++) {
-                            if ((-monthly_electricity_tofrom_grid[m] + (monthly_batt_to_grid[m] + monthly_system_to_grid[m])) != 0)
-                                cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / (-monthly_electricity_tofrom_grid[m] + (monthly_batt_to_grid[m] + monthly_system_to_grid[m])) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //Grid charging cost calculated from monthly energy charges at year a and net annual true ($)
-
-                        }
-                    }
-
-                }
-            }
-            else if (grid_charging_cost_version == 2) { //2-PPA Models besides Single Owner
-                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
-                {
-                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a ($)
-                    for (size_t h = 0; h < n_steps_per_year; h++) {
-                        if (a != 0) {
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(a - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from ppa price ($)
-                        }
-
-                    }
-
-
-                }
-                else //Not lifetime
-                {
-
-                    double ppa_value = cf.at(CF_ppa_price_lcos, a);
-                    for (size_t h = 0; h < n_steps_per_year; h++) {
-                        if (a != 0) { //Not in investment year
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * cf.at(CF_degradation_lcos, a) * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from ppa price ($)
-                        }
-
-                    }
-
-
-                }
-            }
-            else if (grid_charging_cost_version == 3) { //Merchant Plant
-                
-                
-                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
-                {
-                    for (size_t h = 0; h < n_steps_per_year; h++) {
-                        if (a != 0) { //Not in investment year
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(a - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * mp_market_price[(a - 1) * n_steps_per_year + h] / (1000); //Grid charging cost from energy market price ($)
-                        }
-
-                    }
-
-
-                }
-                else //Not Lifetime
-                {
-
-
-                    for (size_t h = 0; h < n_steps_per_year; h++) {
-                        if (a != 0) { //Not in investment year
-                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * cf.at(CF_degradation_lcos, a) * 8760 / n_steps_per_year * mp_market_price[h] / (1000); //Grid charging cost from energy market price ($)
-                        }
-
-                    }
-
-
-                }
-            }
-
-            //cf.at(CF_charging_cost_grid, a) = charged_grid[a] * cf.at(CF_ppa_price, a) / 100; //What is the BTM charge for charging from the grid (do we need to calculate based on utility rates?)
-            //cf.at(CF_charging_cost_grid, a) = charged_grid[a] * 10 / 100; //using 0.10 $/kWh as a placeholder
-            if (cm->as_integer("system_use_lifetime_output") == 1 && a != 0) { //Lifetime
-                cf.at(CF_charging_cost_pv_lcos, a) = charged_pv[a] * lcoe_real_lcos / 100 * pow(1 + inflation_rate, a - 1); //Calculate system charging cost from year a system charged amount ($)
-                cf.at(CF_om_production1_expense_lcos, a) *= lcos_energy_discharged[a];
-                cf.at(CF_energy_discharged_lcos, a) = lcos_energy_discharged[a];
-
-            }
-            else if (cm->as_integer("system_use_lifetime_output") != 1 && a != 0) { //Not lifetime
-                cf.at(CF_charging_cost_pv_lcos, a) = charged_pv[0] * cf.at(CF_degradation_lcos, a) * lcoe_real_lcos / 100 * pow(1 + inflation_rate, a - 1); //Calculate system charging cost from year 1 sytem charged amount ($) (Probably need to account for degradation)
-                cf.at(CF_om_production1_expense_lcos, a) *= lcos_energy_discharged[0] * cf.at(CF_degradation_lcos, a); //Scale om production expense by amount discharged in year 1 ($) account for degradation
-                cf.at(CF_energy_discharged_lcos, a) = lcos_energy_discharged[0] * cf.at(CF_degradation_lcos, a); //Store energy discharged in year 1 to each year of the cash flow
-            }
-            cf.at(CF_annual_cost_lcos_lcos, a) = -cf.at(CF_charging_cost_grid_lcos, a) + //Grid charging cost +
-                -cf.at(CF_charging_cost_pv_lcos, a) + -cf.at(CF_om_fixed1_expense_lcos, a) + //System charging cost + Fixed OM expense + 
-                -cf.at(CF_om_capacity1_expense_lcos, a) + -cf.at(CF_om_production1_expense_lcos, a) + //Capacity based OM expense + Production based OM expense
-                -cf.at(CF_battery_replacement_cost_lcos, a); //Battery replacement expense
-
-        } //End annual iteration
-        cf.at(CF_energy_discharged_lcos, 0) = 0; //Initialize year 0 of energy discharged to 0 for investment year
-        cf.at(CF_annual_cost_lcos_lcos, 0) += -lcos_investment_cost; //add initial investment to year 0
-        double batt_salvage_value_frac = cm->as_double("batt_salvage_percentage") * 0.01; //Battery salvage percentage of initial battery capital cost turned to fraction
-        double lcos_salvage_value = lcos_investment_cost * batt_salvage_value_frac / pow(1 + nom_discount_rate, nyears + 1); //Calculate salvage value of batter only
-        cf.at(CF_salvage_cost_lcos_lcos, nyears) = lcos_salvage_value; //Store salvage value in cash flow
-        cf.at(CF_annual_cost_lcos_lcos, nyears) -= cf.at(CF_salvage_cost_lcos_lcos, nyears); //Add salvage value to overall LCOS cash flow
-        double lcos_denominator = npv(CF_energy_discharged_lcos, nyears, nom_discount_rate, cf); //Find npv of battery energy discharged to use as denominator of LCOS equation (kWh)
-        double lcos_denominator_real = npv(CF_energy_discharged_lcos, nyears, real_discount_rate, cf); //Find npv of battery energy discharged with real discount rate to use as denominator of LCOS equation (kWh)
-        double lcos_numerator = -(npv(CF_annual_cost_lcos_lcos, nyears, nom_discount_rate, cf)) - cf.at(CF_annual_cost_lcos_lcos, 0); //Find npv of battery costs for numerator in LCOS equation ($)
-        cm->assign("npv_annual_costs_lcos", var_data((ssc_number_t)lcos_numerator)); //Store battery cost NPV in outputs
-        save_cf(CF_annual_cost_lcos_lcos, nyears, "cf_annual_cost_lcos", cf, cm); //Store batterys costs in each year in cash flow
-        save_cf(CF_energy_discharged_lcos, nyears, "cf_annual_discharge_lcos", cf, cm); //Store battery energy discharge in each year in cash flow
-        save_cf(CF_charging_cost_grid_lcos, nyears, "cf_charging_cost_grid", cf, cm); //Store grid charging cost in each year in cash flow ($)
-        save_cf(CF_charging_cost_pv_lcos, nyears, "cf_charging_cost_pv", cf, cm); //Store system charging cost in each year in cash flow ($)
-        //save_cf(CF_om_capacity1_expense, nyears, "cf_om_batt_capacity_expense", cf, cm);
-        //save_cf(CF_om_production1_expense, nyears, "cf_om_batt_production_expense", cf, cm);
-        //save_cf(CF_om_fixed1_expense, nyears, "cf_om_batt_fixed_expense", cf, cm);
-        //save_cf(CF_battery_replacement_cost, nyears, "cf_batt_replacement_cost", cf, cm);
-        save_cf(CF_salvage_cost_lcos_lcos, nyears, "cf_salvage_cost_lcos", cf, cm); //Store salvage value cost in each year in cash flow ($)
-        double lcos_nom = lcos_numerator / lcos_denominator * 100.0; // cent/kWh Nominal LCOS
-        double lcos_real = lcos_numerator / lcos_denominator_real * 100.0; // cents/kWh Real LCOS
-        cm->assign("lcos_nom", var_data((ssc_number_t)lcos_nom)); //Store nominal LCOS in outputs
-        cm->assign("lcos_real", var_data((ssc_number_t)lcos_real)); //Store real LCOS in outputs
-        cm->assign("npv_energy_lcos_nom", var_data((ssc_number_t)lcos_denominator)); //Store Nominal LCOS denominator in outputs
-        cm->assign("npv_energy_lcos_real", var_data((ssc_number_t)lcos_denominator_real)); //Store Real LCOS denominator in outputs
-    }
-    /////////////////////////////////////////////////////////////////////////////////////////
-}
+//double npv(int cf_line, int nyears, double rate, util::matrix_t<double> cf)
+//{
+//    //if (rate == -1.0) throw general_error("cannot calculate NPV with discount rate equal to -1.0");
+//    double rr = 1.0;
+//    if (rate != -1.0) rr = 1.0 / (1.0 + rate);
+//    double result = 0;
+//    for (int i = nyears; i > 0; i--)
+//        result = rr * result + cf.at(cf_line, i);
+//
+//    return result * rr;
+//}
+//
+//void escal_or_annual(int cf_line, int nyears, const std::string& variable,
+//    double inflation_rate, double scale,  util::matrix_t<double> cf, compute_module *cm, bool as_rate = true,  double escal = 0.0 )
+//{
+//    size_t count;
+//    ssc_number_t* arrp = cm->as_array(variable, &count);
+//
+//    if (as_rate)
+//    {
+//        if (count == 1)
+//        {
+//            escal = inflation_rate + scale * arrp[0];
+//            for (int i = 0; i < nyears; i++)
+//                cf.at(cf_line, size_t(i) + 1) = pow(1 + escal, i);
+//        }
+//        else
+//        {
+//            for (int i = 0; i < nyears && i < (int)count; i++)
+//                cf.at(cf_line, size_t(i) + 1) = 1 + arrp[i] * scale;
+//        }
+//    }
+//    else
+//    {
+//        if (count == 1)
+//        {
+//            for (int i = 0; i < nyears; i++)
+//                cf.at(cf_line, size_t(i) + 1) = arrp[0] * scale * pow(1 + escal + inflation_rate, i);
+//        }
+//        else
+//        {
+//            for (int i = 0; i < nyears && i < (int)count; i++)
+//                cf.at(cf_line, size_t(i) + 1) = arrp[i] * scale;
+//        }
+//    }
+//}
+//
+//void save_cf(int cf_line, int nyears, const std::string& name, util::matrix_t<double> cf, compute_module *cm)
+//{
+//    ssc_number_t* arrp = cm->allocate(name, size_t(nyears) + 1);
+//    for (int i = 0; i <= nyears; i++)
+//        arrp[i] = (ssc_number_t)cf.at(cf_line, i);
+//}
+//
+//void lcos_calc(compute_module* cm, util::matrix_t<double> cf, int nyears, double nom_discount_rate, double inflation_rate, double lcoe_real, double total_cost, double real_discount_rate, int grid_charging_cost_version, ssc_number_t* ppa_multipliers = { 0 }) {
+//    enum {
+//        CF_battery_replacement_cost_lcos,
+//        CF_ppa_price_lcos,
+//        CF_om_fixed_expense_lcos,
+//        CF_om_production_expense_lcos,
+//        CF_om_capacity_expense_lcos,
+//        CF_om_fixed1_expense_lcos,
+//        CF_om_production1_expense_lcos,
+//        CF_om_capacity1_expense_lcos,
+//        CF_om_fixed2_expense_lcos,
+//        CF_om_production2_expense_lcos,
+//        CF_om_capacity2_expense_lcos,
+//        CF_om_fuel_expense_lcos,
+//        CF_energy_charged_grid_lcos,
+//        CF_energy_charged_pv_lcos,
+//        CF_energy_discharged_lcos,
+//        CF_charging_cost_pv_lcos,
+//        CF_charging_cost_grid_lcos,
+//        CF_om_cost_lcos_lcos,
+//        CF_salvage_cost_lcos_lcos,
+//        CF_investment_cost_lcos_lcos,
+//        CF_annual_cost_lcos_lcos,
+//        CF_util_escal_rate_lcos,
+//        CF_degradation_lcos,
+//
+//    };
+//
+//    if (cm->is_assigned("battery_total_cost_lcos") && cm->as_double("battery_total_cost_lcos") != 0) { //Check that there are non-zero battery costs from the UI
+//        double lcos_investment_cost = cm->as_double("battery_total_cost_lcos"); //Battery capital costs ($), does not include replacement costs
+//        std::vector<double> charged_pv = cm->as_vector_double("batt_annual_charge_from_system"); //Annual kwh charged from PV or generic system (nyears array)
+//        std::vector<double> charged_total = cm->as_vector_double("batt_annual_charge_energy"); //Total annual kwh charged by battery system (nyears array)
+//        std::vector<double> lcos_energy_discharged = cm->as_vector_double("batt_annual_discharge_energy"); //Total annual kwh discharged by battery system (nyears array)
+//        size_t n_grid_to_batt, n_monthly_grid_to_batt, n_monthly_grid_to_load, n_monthly_energy_charge, n_net_annual_true_up; //Preallocate size variables
+//        ssc_number_t* grid_to_batt = cm->as_array("grid_to_batt", &n_grid_to_batt); //Power from grid to battery in kW hourly
+//        ssc_number_t* monthly_grid_to_batt = cm->as_array("monthly_grid_to_batt", &n_monthly_grid_to_batt);
+//        
+//        ssc_number_t* monthly_batt_to_grid = 0;
+//        if (grid_charging_cost_version != 0) //For FOM systems the battery exports to grid rather than to load
+//            monthly_batt_to_grid = cm->as_array("monthly_batt_to_grid", &n_monthly_grid_to_load); //monthly battery exports to grid for 1st year (kWh)
+//        ssc_number_t* monthly_grid_to_load = 0;
+//        if(grid_charging_cost_version == 0) //For BTM systems the battery exports to load
+//            monthly_grid_to_load = cm->as_array("monthly_grid_to_load", &n_monthly_grid_to_load);//monthly battery exports to load for first year (kWh)
+//        
+//
+//        util::matrix_t<double> monthly_energy_charge; //monthly energy charges at 12 month x nyears matrix ($)
+//        util::matrix_t<double> net_annual_true_up; //net annual true up payments as 12 month x nyears matrix ($)
+//        size_t n_steps_per_year = 8760; //Initialize number of timesteps per year (calculated based on lifetime choice and nstep of weather file
+//        //ssc_number_t* monthly_batt_to_grid = as_array("monthly_batt_to_grid", &n_monthly_grid_to_load);
+//        ssc_number_t* monthly_system_to_grid = cm->as_array("monthly_system_to_grid", &n_monthly_grid_to_load); //monthly grid energy bought to satisfy load for first year (kWh)
+//        ssc_number_t* monthly_electricity_tofrom_grid = 0;
+//        bool ppa_purchases = !(cm->is_assigned("en_electricity_rates") && cm->as_number("en_electricity_rates") == 1); //Does the system use ppa purchases in revenue calculations (Single Owner only)
+//        if (!ppa_purchases || (grid_charging_cost_version == 0)) { //Are ppa purchases used in Single Owner or is the case using BTM financial model?
+//            monthly_electricity_tofrom_grid = cm->as_array("year1_monthly_electricity_to_grid", &n_monthly_grid_to_load); //Monthly electricity exported to the grid in the first year (kWh)
+//            monthly_energy_charge = cm->as_matrix("charge_w_sys_ec_ym"); //Use monthly energy charges from utility bill ($)
+//            net_annual_true_up = cm->as_matrix("true_up_credits_ym"); //Use net annual true up payments regardless of billing mode ($)
+//        }
+//        
+//        size_t n_mp_market_price;
+//        ssc_number_t* mp_market_price = 0;
+//        if (grid_charging_cost_version == 3) //Only if financial model is Merchant Plant
+//            mp_market_price = cm->as_array("mp_energy_market_price", &n_mp_market_price); //Energy market price (hourly) for lifetime array to calculate grid charging cost for Merchant Plant ($)
+//
+//        size_t n_degradation;
+//        ssc_number_t* degradation = cm->as_array("degradation", &n_degradation);
+//        if (n_degradation == 1)
+//        {
+//            for (size_t i = 1; i <= nyears; i++) cf.at(CF_degradation_lcos, i) = pow((1.0 - degradation[0] / 100.0), i - 1);
+//        }
+//        else if (n_degradation > 0)
+//        {
+//            for (size_t i = 0; i < nyears && i < (int)n_degradation; i++) cf.at(CF_degradation_lcos, i + 1) = (1.0 - degradation[i] / 100.0);
+//        }
+//
+//        if (cm->as_integer("system_use_lifetime_output") == 1)
+//            n_steps_per_year = n_grid_to_batt / nyears; //Steps per year = Lifetime array size / number years
+//        else
+//            n_steps_per_year = n_grid_to_batt; //If no lifetime, steps per year = first year array size
+//        //std::vector<double> grid_to_batt = as_vector_double("grid_to_batt");
+//        cf.at(CF_charging_cost_grid_lcos, 0) = 0; //Initialize year 0 charging cost to $0
+//        
+//
+//        if (cm->is_assigned("rate_escalation")) //Create rate escalation nyears array with inflation and specified rate escalation %
+//            escal_or_annual(CF_util_escal_rate_lcos, nyears, "rate_escalation", inflation_rate, 0.01, cf, cm, true, 0);
+//        save_cf(CF_util_escal_rate_lcos, nyears, "cf_util_escal_rate", cf, cm);
+//
+//        double capex_lcoe_ratio = 1 / 0.8; //ratio of capex ratio between PV+batt / PV to LCOE ratio PV+batt/ PV (assumed based on table)
+//        double lcoe_real_lcos = lcoe_real * capex_lcoe_ratio * (total_cost - lcos_investment_cost) / total_cost; //cents/kWh
+//        //Using ration of investment cost to lcoe for system+storage and system only to approximate lcoe for system only; Used in PV charging cost calculations
+//       
+//
+//        for (int a = 0; a <= nyears; a++) { //Iterate through nyears of the project
+//
+//
+//            if (grid_charging_cost_version == 0) { //0 - Cashloan, BTM
+//
+//
+//                for (size_t m = 0; m < 12; m++) { //monthly iteration for each year
+//                    if (a != 0 && (monthly_grid_to_load[m] + monthly_grid_to_batt[m]) != 0) {
+//                        //cf.at(CF_charging_cost_grid_month, a) += monthly_grid_to_batt[m] / (monthly_grid_to_batt[m] + monthly_grid_to_load[m]) * monthly_energy_charge[m] * charged_grid[a] / charged_grid[1] * cf.at(CF_util_escal_rate, a);
+//                        cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / (monthly_grid_to_load[m] + monthly_grid_to_batt[m]) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //use the electricity rate data by year (also trueup) //* charged_grid[a] / charged_grid[1] * cf.at(CF_util_escal_rate, a);
+//                    }
+//                }
+//            }
+//            else if (grid_charging_cost_version == 1) { //Single Owner
+//                ppa_purchases = !(cm->is_assigned("en_electricity_rates") && cm->as_number("en_electricity_rates") == 1); //Are ppa purchases used or not?
+//                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
+//                {
+//                    //Lifetime calculations
+//                    //Calculate cost to charge battery from grid (either using ppa price or electricity rates if enabled)
+//                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a
+//                    if (ppa_purchases && a != 0) {
+//                        for (size_t h = 0; h < n_steps_per_year; h++) {
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(size_t(a) - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from PPA price ($)
+//                        }
+//                    }
+//                    else if (!ppa_purchases && a != 0) {
+//                        for (size_t m = 0; m < 12; m++) {
+//                            if (((monthly_batt_to_grid[m] + monthly_system_to_grid[m]) + -monthly_electricity_tofrom_grid[m]) != 0)
+//                                cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / ((monthly_batt_to_grid[m] + monthly_system_to_grid[m]) + -monthly_electricity_tofrom_grid[m]) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //Grid charging cost from monthly energy charges and net annual true up ($
+//
+//                        }
+//                    }
+//
+//
+//                }
+//                else //Not lifetime
+//                {
+//                    //Single year calculations (might not be needed as their is no Single Owner for PVWatts Battery
+//                    //Calculate cost to charge battery from grid (either using ppa price or electricity rates if enabled)
+//                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a ($)
+//
+//                    if (ppa_purchases && a != 0) { //PPA purchases enabled and not in investment year
+//                        for (size_t h = 0; h < 8760; h++) {
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost calculated from PPA price ($)
+//                        }
+//                    }
+//                    else if (!ppa_purchases && a != 0) { //No PPA purchases and not in investment year
+//                        for (size_t m = 0; m < 12; m++) {
+//                            if ((-monthly_electricity_tofrom_grid[m] + (monthly_batt_to_grid[m] + monthly_system_to_grid[m])) != 0)
+//                                cf.at(CF_charging_cost_grid_lcos, a) += monthly_grid_to_batt[m] / (-monthly_electricity_tofrom_grid[m] + (monthly_batt_to_grid[m] + monthly_system_to_grid[m])) * monthly_energy_charge.at(a, m) + net_annual_true_up.at(a, m); //Grid charging cost calculated from monthly energy charges at year a and net annual true ($)
+//
+//                        }
+//                    }
+//
+//                }
+//            }
+//            else if (grid_charging_cost_version == 2) { //2-PPA Models besides Single Owner
+//                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
+//                {
+//                    double ppa_value = cf.at(CF_ppa_price_lcos, a); //PPA price at year a ($)
+//                    for (size_t h = 0; h < n_steps_per_year; h++) {
+//                        if (a != 0) {
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(a - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from ppa price ($)
+//                        }
+//
+//                    }
+//
+//
+//                }
+//                else //Not lifetime
+//                {
+//
+//                    double ppa_value = cf.at(CF_ppa_price_lcos, a);
+//                    for (size_t h = 0; h < n_steps_per_year; h++) {
+//                        if (a != 0) { //Not in investment year
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * cf.at(CF_degradation_lcos, a) * 8760 / n_steps_per_year * ppa_value / 100.0 * ppa_multipliers[h]; //Grid charging cost from ppa price ($)
+//                        }
+//
+//                    }
+//
+//
+//                }
+//            }
+//            else if (grid_charging_cost_version == 3) { //Merchant Plant
+//                
+//                
+//                if (cm->as_integer("system_use_lifetime_output") == 1) //Lifetime
+//                {
+//                    for (size_t h = 0; h < n_steps_per_year; h++) {
+//                        if (a != 0) { //Not in investment year
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[(a - 1) * n_steps_per_year + h] * 8760 / n_steps_per_year * mp_market_price[(a - 1) * n_steps_per_year + h] / (1000); //Grid charging cost from energy market price ($)
+//                        }
+//
+//                    }
+//
+//
+//                }
+//                else //Not Lifetime
+//                {
+//
+//
+//                    for (size_t h = 0; h < n_steps_per_year; h++) {
+//                        if (a != 0) { //Not in investment year
+//                            cf.at(CF_charging_cost_grid_lcos, a) += grid_to_batt[h] * cf.at(CF_degradation_lcos, a) * 8760 / n_steps_per_year * mp_market_price[h] / (1000); //Grid charging cost from energy market price ($)
+//                        }
+//
+//                    }
+//
+//
+//                }
+//            }
+//
+//            //cf.at(CF_charging_cost_grid, a) = charged_grid[a] * cf.at(CF_ppa_price, a) / 100; //What is the BTM charge for charging from the grid (do we need to calculate based on utility rates?)
+//            //cf.at(CF_charging_cost_grid, a) = charged_grid[a] * 10 / 100; //using 0.10 $/kWh as a placeholder
+//            if (cm->as_integer("system_use_lifetime_output") == 1 && a != 0) { //Lifetime
+//                cf.at(CF_charging_cost_pv_lcos, a) = charged_pv[a] * lcoe_real_lcos / 100 * pow(1 + inflation_rate, a - 1); //Calculate system charging cost from year a system charged amount ($)
+//                cf.at(CF_om_production1_expense_lcos, a) *= lcos_energy_discharged[a];
+//                cf.at(CF_energy_discharged_lcos, a) = lcos_energy_discharged[a];
+//
+//            }
+//            else if (cm->as_integer("system_use_lifetime_output") != 1 && a != 0) { //Not lifetime
+//                cf.at(CF_charging_cost_pv_lcos, a) = charged_pv[0] * cf.at(CF_degradation_lcos, a) * lcoe_real_lcos / 100 * pow(1 + inflation_rate, a - 1); //Calculate system charging cost from year 1 sytem charged amount ($) (Probably need to account for degradation)
+//                cf.at(CF_om_production1_expense_lcos, a) *= lcos_energy_discharged[0] * cf.at(CF_degradation_lcos, a); //Scale om production expense by amount discharged in year 1 ($) account for degradation
+//                cf.at(CF_energy_discharged_lcos, a) = lcos_energy_discharged[0] * cf.at(CF_degradation_lcos, a); //Store energy discharged in year 1 to each year of the cash flow
+//            }
+//            cf.at(CF_annual_cost_lcos_lcos, a) = -cf.at(CF_charging_cost_grid_lcos, a) + //Grid charging cost +
+//                -cf.at(CF_charging_cost_pv_lcos, a) + -cf.at(CF_om_fixed1_expense_lcos, a) + //System charging cost + Fixed OM expense + 
+//                -cf.at(CF_om_capacity1_expense_lcos, a) + -cf.at(CF_om_production1_expense_lcos, a) + //Capacity based OM expense + Production based OM expense
+//                -cf.at(CF_battery_replacement_cost_lcos, a); //Battery replacement expense
+//
+//        } //End annual iteration
+//        cf.at(CF_energy_discharged_lcos, 0) = 0; //Initialize year 0 of energy discharged to 0 for investment year
+//        cf.at(CF_annual_cost_lcos_lcos, 0) += -lcos_investment_cost; //add initial investment to year 0
+//        double batt_salvage_value_frac = cm->as_double("batt_salvage_percentage") * 0.01; //Battery salvage percentage of initial battery capital cost turned to fraction
+//        double lcos_salvage_value = lcos_investment_cost * batt_salvage_value_frac / pow(1 + nom_discount_rate, nyears + 1); //Calculate salvage value of batter only
+//        cf.at(CF_salvage_cost_lcos_lcos, nyears) = lcos_salvage_value; //Store salvage value in cash flow
+//        cf.at(CF_annual_cost_lcos_lcos, nyears) -= cf.at(CF_salvage_cost_lcos_lcos, nyears); //Add salvage value to overall LCOS cash flow
+//        double lcos_denominator = npv(CF_energy_discharged_lcos, nyears, nom_discount_rate, cf); //Find npv of battery energy discharged to use as denominator of LCOS equation (kWh)
+//        double lcos_denominator_real = npv(CF_energy_discharged_lcos, nyears, real_discount_rate, cf); //Find npv of battery energy discharged with real discount rate to use as denominator of LCOS equation (kWh)
+//        double lcos_numerator = -(npv(CF_annual_cost_lcos_lcos, nyears, nom_discount_rate, cf)) - cf.at(CF_annual_cost_lcos_lcos, 0); //Find npv of battery costs for numerator in LCOS equation ($)
+//        cm->assign("npv_annual_costs_lcos", var_data((ssc_number_t)lcos_numerator)); //Store battery cost NPV in outputs
+//        save_cf(CF_annual_cost_lcos_lcos, nyears, "cf_annual_cost_lcos", cf, cm); //Store batterys costs in each year in cash flow
+//        save_cf(CF_energy_discharged_lcos, nyears, "cf_annual_discharge_lcos", cf, cm); //Store battery energy discharge in each year in cash flow
+//        save_cf(CF_charging_cost_grid_lcos, nyears, "cf_charging_cost_grid", cf, cm); //Store grid charging cost in each year in cash flow ($)
+//        save_cf(CF_charging_cost_pv_lcos, nyears, "cf_charging_cost_pv", cf, cm); //Store system charging cost in each year in cash flow ($)
+//        //save_cf(CF_om_capacity1_expense, nyears, "cf_om_batt_capacity_expense", cf, cm);
+//        //save_cf(CF_om_production1_expense, nyears, "cf_om_batt_production_expense", cf, cm);
+//        //save_cf(CF_om_fixed1_expense, nyears, "cf_om_batt_fixed_expense", cf, cm);
+//        //save_cf(CF_battery_replacement_cost, nyears, "cf_batt_replacement_cost", cf, cm);
+//        save_cf(CF_salvage_cost_lcos_lcos, nyears, "cf_salvage_cost_lcos", cf, cm); //Store salvage value cost in each year in cash flow ($)
+//        double lcos_nom = lcos_numerator / lcos_denominator * 100.0; // cent/kWh Nominal LCOS
+//        double lcos_real = lcos_numerator / lcos_denominator_real * 100.0; // cents/kWh Real LCOS
+//        cm->assign("lcos_nom", var_data((ssc_number_t)lcos_nom)); //Store nominal LCOS in outputs
+//        cm->assign("lcos_real", var_data((ssc_number_t)lcos_real)); //Store real LCOS in outputs
+//        cm->assign("npv_energy_lcos_nom", var_data((ssc_number_t)lcos_denominator)); //Store Nominal LCOS denominator in outputs
+//        cm->assign("npv_energy_lcos_real", var_data((ssc_number_t)lcos_denominator_real)); //Store Real LCOS denominator in outputs
+//    }
+//    /////////////////////////////////////////////////////////////////////////////////////////
+//}
 
 var_info vtab_equip_reserve[] = {
 { SSC_INPUT,         SSC_NUMBER,    "reserves_interest",                      "Interest on reserves",				                           "%",	 "",					  "Financial Parameters",             "?=1.75",                     "MIN=0,MAX=100",      			"" },
